Add: remove product logic and routes. Update: create product function name. Add: error for resource not found

+++ src/use-cases/remove-product.ts
@@ -1,4 +1,5 @@
 import { UsersProductsRepository } from '@/repositories/interfaces/users-products-repository';
+import { ResourceNotFoundError } from './errors/resource-not-found-error';
 
 interface RemoveProductUseCaseRequest {
 	productId: string;
@@ -8,6 +9,12 @@ export class RemoveProductUseCase {
 	constructor(private usersProductsRepository: UsersProductsRepository) {}
 
 	async execute({ productId }: RemoveProductUseCaseRequest): Promise<void> {
+		const product = await this.usersProductsRepository.findByProductId(productId);
+
+		if (!product) {
+			throw new ResourceNotFoundError();
+		}
+
 		await this.usersProductsRepository.delete(productId);
 	}
 }
